[FEAT](ai): implemented fork
