+++ src/AI/agent/decisionManager.py
@@ -18,13 +18,14 @@ class DecisionManager:
       "TakeOneFoodHere": behaviors.TakeOneFoodHereBehavior(agent),
       "DropEveryMinerals": behaviors.DropEveryMineralsBehavior(agent),
       "DropAllFood": behaviors.DropAllFoodBehavior(agent),
+      "Fork": behaviors.ForkBehavior(agent),
       "None": behaviors.NoActionBehavior(agent),
       "": behaviors.NoActionBehavior(agent),
     }
 
 
     self.decisions = {
-      "collecting": {"miner": ["BigDyson"], "fighter": ["FoodDyson"]},
+      "collecting": {"miner": ["Fork", "BigDyson"], "fighter": ["FoodDyson"]},
       "rallying": {"miner": ["JoinTeamMates", "TakeAllFoodHere"], "fighter": ["JoinTeamMates"]},
       "setting": {"miner": ["DropEveryMinerals"], "fighter": ["DropAllFood"]},
       "upgrading": {"miner": ["Upgrade", "TakeOneFoodHere", "TakeOneFoodHere"], "fighter": ["FoodBigDyson"]}
