+++ src/AI/main.py
@@ -1,37 +1,9 @@
 #!/usr/bin/env python3
-import sys
-import os
+
 import signal
-from time import sleep
+import utils.mutliprocessing as mutliprocessing
 from agent.agent import Agent
 from parser.paringArgsClass import parseArgs
-
-agents = []
-
-def run_agent(ip, port, team, agent_id, performance):
-  try:
-    agent = Agent(ip, port, team, agent_id, performance)
-    agents.append(agent)
-    agent.start()
-  except Exception as e:
-    print(f"Agent {agent_id} failed: {e}")
-    sys.exit(1)
-
-def signal_handler(sig, frame):
-  print("Received termination signal. Stopping all agents...")
-  for agent in agents:
-    try:
-      agent.stop()
-    except Exception as e:
-      print(f"Error stopping agent: {e}")
-
-  for pid in child_pids:
-    try:
-      os.kill(pid, signal.SIGTERM)
-      print(f"Sent SIGTERM to process {pid}")
-    except ProcessLookupError:
-      pass
-  sys.exit(0)
 
 if __name__ == "__main__":
   args = parseArgs()
@@ -45,27 +17,13 @@ if __name__ == "__main__":
   if performance:
     print("Performance mode enabled")
 
-  child_pids = []
+  signal.signal(signal.SIGINT, mutliprocessing.signal_handler)
+  signal.signal(signal.SIGTERM, mutliprocessing.signal_handler)
 
-  signal.signal(signal.SIGINT, signal_handler)
-  signal.signal(signal.SIGTERM, signal_handler)
-
-  for i in range(num_agents):
-    pid = os.fork()
-
-    if pid == 0:
-      run_agent(ip, port, team, i, performance)
-      sys.exit(0)
-    else:
-      child_pids.append(pid)
-      print(f"Started agent {i} with PID {pid}")
-      sleep(0.01)
+  mutliprocessing.run_agent(ip, port, team, 0, performance)
 
   print("IA started successfully!")
 
-  try:
-    for pid in child_pids:
-      os.waitpid(pid, 0)
-  except KeyboardInterrupt:
-    pass
+  mutliprocessing.wait_childs()
+
   print("All agents completed")
